test: add usecase unittest
